+++ src/main/java/org/jusecase/scenegraph/math/Vector2.java
@@ -1,13 +1,13 @@
 package org.jusecase.scenegraph.math;
 
 public final class Vector2 {
-    public double x;
-    public double y;
+    public float x;
+    public float y;
 
     public Vector2() {
     }
 
-    public Vector2(double x, double y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
