+++ src/main/java/org/jusecase/scenegraph/texture/TexCoords.java
@@ -3,13 +3,13 @@ package org.jusecase.scenegraph.texture;
 public class TexCoords {
     public static final TexCoords DEFAULT = new TexCoords(0, 1, 1, 0);
 
-    public final double left;
-    public final double bottom;
-    public final double right;
-    public final double top;
+    public final float left;
+    public final float bottom;
+    public final float right;
+    public final float top;
 
 
-    public TexCoords(double left, double bottom, double right, double top) {
+    public TexCoords(float left, float bottom, float right, float top) {
         this.left = left;
         this.bottom = bottom;
         this.right = right;
